server/requirements.txt
mysql-connector-python
fastapi
uvicorn
dotenv



